bug api

+++ projet_Fulbank/AchatCrypto1.cs
@@ -53,7 +53,7 @@ namespace projet_Fulbank
         {
             try
             {
-                textBox2.Text = (AppelHTTPS.GetAmountCrypto(typecrypto.SelectedItem.ToString(), float.Parse(textBox1.Text))).ToString();
+                textBox2.Text = (AppelHTTPS.GetAmountCrypto(typecrypto.SelectedItem.ToString(), double.Parse(textBox1.Text))).ToString();
             }
             catch (Exception ex)
             {
